exceptions may pertain to the future
E.g., we plan to use a non-FOSS tool internally, but our policies say
we prefer to look for FOSS alternative even if we have to self-host.

Exceptions should be stated in terms of the risks we accept and known
and/or predictable specific circumstate but should not provide an
explicit basis for naming an individual not otherwise named except in
the need to establish documentation of or otherwise handle abuse.

Or, sez me.